add / i2i に関する操作をする機能をもつクラスを追加。
